requirements.txt
flask>=3.0
PyMuPDF>=1.24
paddlepaddle
paddleocr>=2.7
Pillow>=10.0
numpy
